+++ sat/system/node.py
@@ -29,6 +29,9 @@ from sat.system.constants import CAB_TYPE_C, CAB_TYPE_S, NODE_TYPE
 from sat.system.drive import Drive
 from sat.system.field import ComponentField
 from sat.system.memory_module import MemoryModule
+from sat.system.node_accel import NodeAccel
+from sat.system.node_accel_riser import NodeAccelRiser
+from sat.system.node_hsn_nic import NodeHsnNic
 from sat.system.processor import Processor
 from sat.util import bytes_to_gib
 
@@ -54,6 +57,9 @@ class Node(BaseComponent):
         ComponentField('Processor Count'),
         ComponentField('Processor Manufacturer', summarizable=True),
         ComponentField('Processor Model', summarizable=True),
+        ComponentField('Accelerator Count', summarizable=True),
+        ComponentField('Accelerator Riser Count', summarizable=True),
+        ComponentField('HSN NIC Count', summarizable=True),
         ComponentField('Drive Count', summarizable=True),
         ComponentField('Total Drive Capacity (GiB)', summarizable=True),
         ComponentField('BIOS Version'),
@@ -75,11 +81,17 @@ class Node(BaseComponent):
         # are children of this node.
         self.memory_modules = {}
         self.processors = {}
+        self.node_accels = {}
+        self.node_accel_risers = {}
+        self.node_hsn_nics = {}
         self.drives = {}
 
         self.children_by_type = {
             MemoryModule: self.memory_modules,
             Processor: self.processors,
+            NodeAccel: self.node_accels,
+            NodeAccelRiser: self.node_accel_risers,
+            NodeHsnNic: self.node_hsn_nics,
             Drive: self.drives
         }
 
@@ -139,6 +151,21 @@ class Node(BaseComponent):
     def memory_module_count(self):
         """int: The number of memory modules this node has."""
         return len(self.memory_modules)
+
+    @cached_property
+    def accelerator_count(self):
+        """int: The number of node_accels this node has."""
+        return len(self.node_accels)
+
+    @cached_property
+    def accelerator_riser_count(self):
+        """int: The number of node_accel_risers this node has."""
+        return len(self.node_accel_risers)
+
+    @cached_property
+    def hsn_nic_count(self):
+        """int: The number of node_hsn_nics this node has."""
+        return len(self.node_hsn_nics)
 
     @cached_property
     def drive_count(self):