+++ sat/system/system.py
@@ -34,7 +34,11 @@ from sat.system.hsn_board import HSNBoard
 from sat.system.memory_module import MemoryModule
 from sat.system.node import Node
 from sat.system.node_enclosure import NodeEnclosure
+from sat.system.node_enclosure_power_supply import NodeEnclosurePowerSupply
 from sat.system.processor import Processor
+from sat.system.node_accel import NodeAccel
+from sat.system.node_accel_riser import NodeAccelRiser
+from sat.system.node_hsn_nic import NodeHsnNic
 from sat.system.router_module import RouterModule
 from sat.xname import XName
 
@@ -63,7 +67,11 @@ class System:
             MemoryModule: {},
             Node: {},
             NodeEnclosure: {},
+            NodeEnclosurePowerSupply: {},
             Processor: {},
+            NodeAccel: {},
+            NodeAccelRiser: {},
+            NodeHsnNic: {},
             RouterModule: {}
         }