+++ tests/system/test_node.py
@@ -32,6 +32,9 @@ from sat.system.processor import Processor
 from tests.system.component_data import get_component_raw_data, CHASSIS_XNAME, NODE_XNAME
 
 DEFAULT_PROCESSOR_COUNT = 2
+DEFAULT_NODE_ACCEL_COUNT = 1
+DEFAULT_NODE_ACCEL_RISER_COUNT = 1
+DEFAULT_NODE_HSN_NIC_COUNT = 2
 DEFAULT_BIOS_VERSION = '2019.11.b'
 
 
@@ -90,6 +93,39 @@ class TestNode(unittest.TestCase):
         self.node.processors = {'{}p{}'.format(NODE_XNAME, index): proc
                                 for index, proc in enumerate(self.fake_processors)}
 
+    def add_fake_node_accels(self):
+        """Add some fake node accelerators to `self.node`."""
+        self.fake_node_accels = []
+        for _ in range(DEFAULT_NODE_ACCEL_COUNT):
+            fake_node_accel = mock.Mock()
+            self.fake_node_accels.append(fake_node_accel)
+        # Set the node accelerators in the node with some xnames beneath the node, not
+        # that it should matter.
+        self.node.node_accels = {'{}a{}'.format(NODE_XNAME, index): node_accel
+                                 for index, node_accel in enumerate(self.fake_node_accels)}
+
+    def add_fake_node_accel_risers(self):
+        """Add some fake node accelerator risers to `self.node`."""
+        self.fake_node_accel_risers = []
+        for _ in range(DEFAULT_NODE_ACCEL_RISER_COUNT):
+            fake_node_accel_riser = mock.Mock()
+            self.fake_node_accel_risers.append(fake_node_accel_riser)
+        # Set the node accelerator risers in the node with some xnames beneath the node, not
+        # that it should matter.
+        self.node.node_accel_risers = {'{}r{}'.format(NODE_XNAME, index): node_accel_riser
+                                       for index, node_accel_riser in enumerate(self.fake_node_accel_risers)}
+
+    def add_fake_node_hsn_nics(self):
+        """Add some fake node HSN NICs to `self.node`."""
+        self.fake_node_hsn_nics = []
+        for _ in range(DEFAULT_NODE_HSN_NIC_COUNT):
+            fake_node_hsn_nic = mock.Mock()
+            self.fake_node_hsn_nics.append(fake_node_hsn_nic)
+        # Set the node HSN NICs in the node with some xnames beneath the node, not
+        # that it should matter.
+        self.node.node_hsn_nics = {'{}h{}'.format(NODE_XNAME, index): node_hsn_nic
+                                   for index, node_hsn_nic in enumerate(self.fake_node_hsn_nics)}
+
     def add_fake_drives(self):
         """Add some fake drives to `self.node`."""
         self.fake_drives = []
@@ -139,6 +175,21 @@ class TestNode(unittest.TestCase):
         """Test the processor_count property."""
         self.add_fake_processors()
         self.assertEqual(DEFAULT_PROCESSOR_COUNT, self.node.processor_count)
+
+    def test_accelerator_count(self):
+        """Test the accelerator_count property."""
+        self.add_fake_node_accels()
+        self.assertEqual(DEFAULT_NODE_ACCEL_COUNT, self.node.accelerator_count)
+
+    def test_accelerator_riser_count(self):
+        """Test the accelerator_riser_count property."""
+        self.add_fake_node_accel_risers()
+        self.assertEqual(DEFAULT_NODE_ACCEL_RISER_COUNT, self.node.accelerator_riser_count)
+
+    def test_hsn_nic_count(self):
+        """Test the hsn_nic_count property."""
+        self.add_fake_node_hsn_nics()
+        self.assertEqual(DEFAULT_NODE_HSN_NIC_COUNT, self.node.hsn_nic_count)
 
     def test_memory_type(self):
         """Test the memory_type property."""